Trivial changes

+++ src/parser/ctors.h
@@ -9,10 +9,10 @@ extern Node* make_oper(Arena* arena, OperKind oper, Node* lhs, Node* rhs);
 extern Node* make_unary(Arena* arena, NodeKind kind, Node* value);
 extern Node* make_basic_value(Arena* arena, Node* type, StrView view);
 extern Node* make_str_value(Arena* arena, StrView view);
-extern Node* make_numeric_value(Arena* arena, Node* type, u64 num);
+extern Node* make_numeric_value(Arena* arena, Node* type, usize num);
 extern Node* make_pointer_value(Arena* arena, Node* type, Node* value);
 extern Node* make_basic_type(Arena* arena, TypeKind kind);
-extern Node* make_numeric_type(Arena* arena, TypeKind kind, i32 width);
+extern Node* make_numeric_type(Arena* arena, TypeKind kind, usize width);
 extern Node* make_pointer_type(Arena* arena, Node* value);
 extern Node* make_array_type(Arena* arena, Node* type, usize size);
 extern Node* make_declaration(
